# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend API for Nillumbik Shire project.",
        "title": "Nillubim Shire API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/",
    "paths": {
        "/observations": {
            "get": {
                "description": "List observations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "observation"
                ],
                "summary": "List observations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/observation.Observation"
                            }
                        }
                    }
                }
            }
        },
        "/observations/{id}": {
            "get": {
                "description": "Get the detail of an observation by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "observation"
                ],
                "summary": "Get Observation Detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the observation",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/observation.Observation"
                        }
                    }
                }
            }
        },
        "/sites": {
            "get": {
                "description": "List sites",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "site"
                ],
                "summary": "List sites",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Site"
                            }
                        }
                    }
                }
            }
        },
        "/sites/{code}": {
            "get": {
                "description": "Get the detail of a site by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "site"
                ],
                "summary": "Get Site Detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Code of the site",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Site"
                        }
                    }
                }
            }
        },
        "/species": {
            "get": {
                "description": "list all species",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "species"
                ],
                "summary": "List species",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Species"
                            }
                        }
                    }
                }
            }
        },
        "/species/{id}": {
            "get": {
                "description": "Get species detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "species"
                ],
                "summary": "Get species detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of the species",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Species"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "db.ForestType": {
            "type": "string",
            "enum": [
                "dry",
                "wet"
            ],
            "x-enum-varnames": [
                "ForestTypeDry",
                "ForestTypeWet"
            ]
        },
        "db.ObservationMethod": {
            "type": "string",
            "enum": [
                "audio",
                "camera",
                "observed"
            ],
            "x-enum-varnames": [
                "ObservationMethodAudio",
                "ObservationMethodCamera",
                "ObservationMethodObserved"
            ]
        },
        "db.Site": {
            "type": "object",
            "properties": {
                "block": {
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "forest": {
                    "$ref": "#/definitions/db.ForestType"
                },
                "id": {
                    "type": "integer"
                },
                "location": {},
                "name": {
                    "type": "string"
                },
                "tenure": {
                    "$ref": "#/definitions/db.TenureType"
                }
            }
        },
        "db.Species": {
            "type": "object",
            "properties": {
                "common_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "native": {
                    "type": "boolean"
                },
                "scientific_name": {
                    "type": "string"
                },
                "taxa": {
                    "$ref": "#/definitions/db.Taxa"
                }
            }
        },
        "db.Taxa": {
            "type": "string",
            "enum": [
                "bird",
                "mammal",
                "reptile"
            ],
            "x-enum-varnames": [
                "TaxaBird",
                "TaxaMammal",
                "TaxaReptile"
            ]
        },
        "db.TenureType": {
            "type": "string",
            "enum": [
                "public",
                "private"
            ],
            "x-enum-varnames": [
                "TenureTypePublic",
                "TenureTypePrivate"
            ]
        },
        "observation.Observation": {
            "type": "object",
            "properties": {
                "appearance_time": {
                    "type": "object",
                    "properties": {
                        "end": {
                            "type": "integer"
                        },
                        "start": {
                            "type": "integer"
                        }
                    }
                },
                "confidence": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "indicator": {
                    "type": "boolean"
                },
                "method": {
                    "$ref": "#/definitions/db.ObservationMethod"
                },
                "narrative": {
                    "type": "string"
                },
                "reportable": {
                    "type": "boolean"
                },
                "site_id": {
                    "type": "integer"
                },
                "species_id": {
                    "type": "integer"
                },
                "temperature": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}